Get the async/await multiplexer working.
Previous attempts involved using a futures::lock::BiLock to share the TCP
connection, but because we aquired the lock, read from the stream as seperate
steps, it'd be entirely possible for the reader side of the connection to
block the writer.

Conversely, the SplitSink/Stream from `StreamExt::split()` use
`BiLock::poll_lock()` from within a `ready!()` macro in their `poll()` method.
The lesson being be careful of your lock scope, folks.